requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
blessed==1.20.0; python_version >= '2.7'
boto3==1.34.99; python_version >= '3.8'
botocore==1.34.99; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
chalice==1.31.0
chalice-local==0.1.1
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
idna==3.7; python_version >= '3.5'
inquirer==2.10.1; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
-e .
localstack-client==2.5
pillow==10.3.0
pip==23.3.2; python_version >= '3.7'
pydantic==2.7.1; python_version >= '3.8'
pydantic-core==2.18.2; python_version >= '3.8'
pydantic-settings==2.2.1; python_version >= '3.8'
pypdf2[image]==3.0.1; python_version >= '3.6'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-editor==1.0.4
pyyaml==6.0.1; python_version >= '3.6'
readchar==4.0.6; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
s3transfer==0.10.1; python_version >= '3.8'
setuptools==69.5.1; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.11.0; python_version >= '3.8'
urllib3==2.2.1; python_version >= '3.8'
wcwidth==0.2.13
wheel==0.43.0; python_version >= '3.8'
